+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created Recent Items panel, linked to Browse panel [#1701](https://github.com/rokwire/illinois-app/issues/1701).
 ### Changed
 - Cleaned up recent items handling [#1708](https://github.com/rokwire/illinois-app/issues/1708).
+- Updated Wellness health rings (Work in progress)  [#1692](https://github.com/rokwire/illinois-app/issues/1692).
 
 ## [4.0.27] - 2022-06-10
 ### Added